Restructure makefile variable definition
The existing definition of FWVERSION has the toplevel
structure of an if-then-else, each terminal of which is an
assignment. With BSD makes, this is problematic (tested with
OpenBSD 6.5). To make such makes happe, the definition of
FWVERSION should have a traditional

  VAR = VAL

shape. In this commit, we achieve that by "shelling out",
putting the if-then-else logic into a $(shell ...)
expression.

Tested on OpenBSD 6.5 and macOS Mojave, both with its
built-in make and GNU Make.