Explode CI job configuration for Gitlab pipelines
To obtained more fine grained results, I have exploded the pipeline by
splitting jobs compiling each of the versions of racket (cgc, 3m and
cs). I have then used each of these to run the respective tests, and
each test collection have its own job as well. This is a refactoring
to obtain better results and part of the long term plan of being able
to generate a user/contributor readable at-a-glance dashboard for the
racket ci system.

Unfortunately we are hitting the capabilities of gitlab, due to its
inability to represent pipeline dags. Although Gitlab CI maintainers
are aware of this a solution within the platform is being constantly
delayed - we might have to look elsewhere or if we are really
motivated, develop an in-Racket CI system.