Small cleanup of hash-ref code
Instead of having different kinds of hash implement the
default-value-or-thunk-on-failure protocol, let's move
that to hash.ss and make the internal functions just
accept a default value (with no special handling for thunks).

I think this makes the code a bit simpler, but it also should
make it easier to abstract over the hash-ref code, which,
in turn, should make it easier to add hash-ref-pair and
hash-ref-key. (Or so I hope.)