Relicense this repository under Apache 2.0/MIT.
Note: the traditional Racket executable continues to incorporate
LGPLv3 code and thus modifications to it must be released under
that license. However, all Racket code in this repository, as well
as the Racket-on-Chez excutable, are more permissively licensed.

Thanks in particular to @otherjoel and @zyrolasting for their work
on this project.

(cherry picked from commit 7783aee9a733065458678f34b8538082e165af25)