Add comment to CI step (non-functional)
@samth requested a clarification of the condition under #2959 
Hopefully this comment makes it clear that we require the condition in order not to tarball and upload the clang artifact on Linux.

The choice not to test both the gcc and clang builds was made to save time and because we are not testing the compiler itself, but racket. 
Hopefully by the time we get these production compilers, they are generating correct code.